strace: add new `--uid`, `--gid` options for static linking
When building strace statically linked using the "--username"
argument will lead to strange errors. The reason is that the
calls to getpwnam() and initgroups() will fail because glibc
will still try to dlopen() the libnss modules which then
fails in unexpected ways.

The `--username` argument is only needed for uid,gid and
supplementary groups setup. So this commit provides a way
to pass uid,gid via the commandline and it will not init
supplementary groups (but it could be extended for this).

This does allow to use a statically linked strace.